requirements.txt
flask
torch
torchaudio
groq
langchain
langchain-groq
chatterbox-tts
requests
python-dotenv